public/code/requirements.txt
Flask==3.0.3
google-cloud-storage